+++ Mods ChoGGi/Construction Show Drone Grid/Code/Script.lua
@@ -1,3 +1,5 @@
+-- See LICENSE for terms
+
 
 -- local whatever globals we call
 local ShowHexRanges = ShowHexRanges